fix: type

+++ src/resume_screener/models/aggregation.py
@@ -9,7 +9,7 @@ class AgentSignal(BaseModel):
     """Signal from an individual agent in the pipeline."""
 
     agent_name: str
-    score: int | None = Field(default=None, ge=0, le=100)
+    score: float | None = Field(default=None, ge=0, le=100)
     confidence: Literal["high", "medium", "low"]
     key_findings: list[str] = Field(default_factory=list)
     concerns: list[str] = Field(default_factory=list)
@@ -19,12 +19,12 @@ class FinalRecommendation(BaseModel):
     """Final hiring recommendation aggregated from all agents."""
 
     recommendation: Literal["Strong Hire", "Hire", "Maybe", "No Hire", "Strong No Hire"]
-    overall_score: int = Field(ge=0, le=100)
+    overall_score: float = Field(ge=0, le=100)
     confidence: Literal["high", "medium", "low"]
     candidate_name: str
     candidate_email: str | None = None
-    skill_match_score: int = Field(ge=0, le=100)
-    experience_fit_score: int = Field(ge=0, le=100)
+    skill_match_score: float = Field(ge=0, le=100)
+    experience_fit_score: float = Field(ge=0, le=100)
     agent_signals: list[AgentSignal] = Field(default_factory=list)
     summary: str = Field(description="Executive summary of the recommendation")
     strengths: list[str] = Field(default_factory=list)
